requirements.txt
Flask
azure-cognitiveservices-vision-computervision
openai
requests